+++ src/main.js
@@ -1,12 +1,14 @@
-import Vue from "vue";
-import App from "./App.vue";
-import Router from "./router";
+import Vue from 'vue';
+import App from './App.vue';
+import Router from './router';
 import currency from 'v-currency-field';
-import store from "./store";
-import vuetify from "@/plugins/vuetify";
-import VueSweetalert2 from "vue-sweetalert2";
-import "sweetalert2/dist/sweetalert2.min.css";
+import store from './store';
+import vuetify from '@/plugins/vuetify';
+import VueSweetalert2 from 'vue-sweetalert2';
+import 'sweetalert2/dist/sweetalert2.min.css';
+import Vuelidate from 'vuelidate';
 
+Vue.use(Vuelidate);
 Vue.use(VueSweetalert2);
 Vue.use(currency);
 Vue.config.productionTip = false;
@@ -16,4 +18,4 @@ new Vue({
   vuetify,
   store,
   render: (h) => h(App),
-}).$mount("#app");
+}).$mount('#app');