+++ CHANGELOG.md
@@ -157,6 +157,11 @@
 - Modul Tambah Kehadiran
     - tambah status masuk / tidak masuk karyawan
 
+# 0.0.1 (Update 16/02/2023)
+- Modul Payslip
+    - tambah dialog konfirmasi untuk cetak payslip
+    
+
 # 0.0.1 (Update 22/02/2023)
 - Modul Payslip Produksi
     - Solve Payslip update bon auto refresh